Move panic handler to executor crate
