+++ ic-cdk-executor/Cargo.toml
@@ -15,3 +15,4 @@ links = "ic-cdk async executor, see https://github.com/dfinity/cdk-rs/blob/links
 
 [dependencies]
 slotmap.workspace = true
+ic0.workspace = true
