--- ic-cdk/src/printer.rs
@@ -1,28 +1,0 @@
-use crate::api;
-use std::panic;
-
-/// Sets a custom panic hook, uses debug.trace
-pub fn set_panic_hook() {
-    panic::set_hook(Box::new(|info| {
-        let file = info.location().unwrap().file();
-        let line = info.location().unwrap().line();
-        let col = info.location().unwrap().column();
-
-        let msg = match info.payload().downcast_ref::<&'static str>() {
-            Some(s) => *s,
-            None => match info.payload().downcast_ref::<String>() {
-                Some(s) => &s[..],
-                None => "Box<Any>",
-            },
-        };
-
-        let err_info = format!("Panicked at '{msg}', {file}:{line}:{col}");
-        api::debug_print(&err_info);
-        api::trap(&err_info);
-    }));
-}
-
-/// Sets stdout, stderr, and a custom panic hook
-pub fn hook() {
-    set_panic_hook();
-}
